+++ src/parser/text/plaintext.ts
@@ -1,18 +1,16 @@
 ﻿import { Result } from '../../combinator/parser';
-import { PlainTextParser, Zalgo } from '../text';
-import { zalgo } from './zalgo/plaintext';
+import { PlainTextParser } from '../text';
 
-const separator = /`|<\/code>|\n|[\u0300-\u036F]/i;
+const separator = /`|<\/code>|\n/i;
 
-export const plaintext: PlainTextParser = function (source: string): Result<Text, [Zalgo.ZalgoPlainTextParser]> {
+export const plaintext: PlainTextParser = function (source: string): Result<Text, [never]> {
   if (source.length === 0) return;
   const i = source.search(separator);
   switch (i) {
     case -1:
       return [[document.createTextNode(source)], ''];
     case 0:
-      return zalgo(source)
-          || [[document.createTextNode(source.slice(0, 1))], source.slice(1)];
+      return [[document.createTextNode(source.slice(0, 1))], source.slice(1)];
     default:
       return [[document.createTextNode(source.slice(0, i))], source.slice(i)];
   }