+++ src/parser/text/text.test.ts
@@ -63,12 +63,6 @@ describe('Unit: parser/text', () => {
       assert.deepStrictEqual(inspect(parser('。\n。')), [['。', '。'], '']);
     });
 
-    it('zalgo', () => {
-      assert.deepStrictEqual(inspect(parser('\u0300\u036F')), [['\u0300'], '']);
-      assert.deepStrictEqual(inspect(parser('\\\u0300\\\u036F')), [['\u0300'], '']);
-      assert.deepStrictEqual(inspect(parser('a\u0300\u036F')), [['a', '\u0300'], '']);
-    });
-
     it('uri', () => {
       assert.deepStrictEqual(inspect(parser('http:')), [['http', ':'], '']);
       assert.deepStrictEqual(inspect(parser('https:')), [['https', ':'], '']);