+++ src/parser/text/text.ts
@@ -1,11 +1,10 @@
 import { Result } from '../../combinator/parser';
-import { TextParser, Zalgo } from '../text';
-import { zalgo } from './zalgo/text';
+import { TextParser } from '../text';
 
 const separator = /[^0-9a-zA-Z\u0080-\uFFFF]|[\u0300-\u036F]|(?:[0-9a-zA-Z][!?]*h|\?h|[0-9a-gi-zA-Z!?])ttps?:|[0-9a-zA-Z][!?]*#\S|[0-9a-zA-Z@]?@[0-9a-zA-Z]|[、。]/u;
 const linebreaks = /^(?:(?:\\?\s)*?\\?\n)+/;
 
-export const text: TextParser = function (source: string): Result<HTMLBRElement | Text, [Zalgo.ZalgoTextParser]> {
+export const text: TextParser = function (source: string): Result<HTMLBRElement | Text, [never]> {
   if (source.length === 0) return;
   const i = source.search(separator);
   switch (i) {
@@ -18,8 +17,7 @@ export const text: TextParser = function (source: string): Result<HTMLBRElement 
             case '\n':
               return [[document.createElement('br')], source.replace(linebreaks, '')];
             default:
-              return zalgo(source)
-                  || [[document.createTextNode(source.slice(1, 2))], source.slice(2)];
+              return [[document.createTextNode(source.slice(1, 2))], source.slice(2)];
           }
         case '、':
         case '。':
@@ -32,8 +30,7 @@ export const text: TextParser = function (source: string): Result<HTMLBRElement 
         case '\n':
           return [[document.createTextNode(' ')], source.slice(1)];
         default:
-          return zalgo(source)
-              || [[document.createTextNode(source.slice(0, 1))], source.slice(1)];
+          return [[document.createTextNode(source.slice(0, 1))], source.slice(1)];
       }
     default:
       return [[document.createTextNode(source.slice(0, i))], source.slice(i)];